Initial checkin. Dan O'Neil - 3/17/2011 - Added constructor overloads in Policy Server, 	to allow for specific ports and IP addresses; 	Made appropriate modifications to Policy Server to work with 	a Flash client.
SocketsLight was originally created by Alexey Zakharov, under the GNU GPL v3 or later.

SocketsLight Framework is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

SocketsLight Framework is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with Foobar.  If not, see http://www.gnu.org/license.